refactoring lines
