+++ api/cases/handlers.js
@@ -188,10 +188,14 @@ module.exports = (server) => {
                 query,
                 request.auth.credentials.user,
                 (err, result) => {
-                  if (err) return reply(replyHelper.constructErrorResponse(err)).code(422)
-                  return reply(
-                      constructCasesResponse(result,request)
-                  ).code(200)
+                if (err) return reply(replyHelper.constructErrorResponse(err)).code(422)
+                const jsonXls = json2xls(result);
+                const fileName = `Data-Kasus-${fullName}-${moment().format("YYYY-MM-DD-HH-mm")}.xlsx`
+                fs.writeFileSync(fileName, jsonXls, 'binary');
+                const xlsx = fs.readFileSync(fileName)
+                reply(xlsx)
+                .header('Content-Disposition', 'attachment; filename='+fileName);
+                return fs.unlinkSync(fileName);
             })
         },
         /**
